backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
python-dotenv
pandas 
openpyxl 
sqlalchemy 
psycopg2-binary 
requests